Tests fixed and pass
