+++ src/spatial_discretization/tesselation_1d.py
