+++ src/test_all.py
@@ -1,6 +1,7 @@
 import unittest
 
 import tests.test_maxwell
+import tests.test_time_integrator
 
 if __name__ == '__main__':
     unittest.main()
