+++ src/time_integrator.py
@@ -1,6 +1,33 @@
-import numpy as np
+import time
 import math
-from time_integrator import TimeIntegrator
+import numpy as np
+
+
+
+class TimeIntegrator:
+    """
+        Time integrators perform the evolution of a set of equations of the form
+            d{f} / dt = [A] {f}(t) + {g}(t)
+        with {f} being a vector of fields, [A] is a matrix of coefficients, and
+        {g} is a forcing function. 
+    """
+    __timeStepPrint = 10
+
+    def start_timer(self):
+        self.tic = time.time()
+
+    def print_progress(self, n, number_of_steps):
+        if n % self.__timeStepPrint == 0 or n+1 == number_of_steps:
+            remaining = (time.time() - self.tic) *  (number_of_steps-n) / (n+1)
+            min = math.floor(remaining / 60.0)
+            sec = remaining % 60.0
+            print("  Step: %6d of %6d. Remaining: %2.0f:%02.0f"% \
+                (n, number_of_steps-1, min, sec))
+
+
+    def print_cpu_time(self):
+        print("    CPU Time: %f [s]" % (time.time() - self.tic))
+
 
 class LSERK4(TimeIntegrator):
     a = np.array([                             0.0, \
@@ -24,7 +51,7 @@ class LSERK4(TimeIntegrator):
         self.residue = np.zeros(equation.vars.size)
 
     def integrate(self, dt, number_of_steps):
-        TimeIntegrator.start_timer()
+        # self.start_timer()
         
         t = 0.0
         for n in range(number_of_steps):
@@ -34,6 +61,6 @@ class LSERK4(TimeIntegrator):
                 self.equation.vars += self.b[i] * self.residue
 
             t += dt
-            TimeIntegrator.print_progress(n, number_of_steps)
+            # self.print_progress(n, number_of_steps)
 
-        TimeIntegrator.print_cpu_time()
+        # self.print_cpu_time()
